Remove test_output before creating directories

+++ test_preprocess.py
@@ -9,11 +9,13 @@ class TestPreprocess(unittest.TestCase):
 	OUTPUT_DIR = 'test_output'
 	TIMES = [20150208093000, 20171113141130, 20171113192200]
 
-	def tearDown(self):
-		rmtree(self.OUTPUT_DIR)
+	@classmethod
+	def setUpClass(self):
+		if os.path.isdir(self.OUTPUT_DIR):
+			rmtree(self.OUTPUT_DIR)
+		create_dirs(self.TIMES, self.OUTPUT_DIR)
 
 	def test_creates_directories(self):
-		create_dirs(self.TIMES, self.OUTPUT_DIR)
 		self.assertTrue(os.path.isdir(self.OUTPUT_DIR + '/simpleimage/2015/0208'))
 		self.assertTrue(os.path.isdir(self.OUTPUT_DIR + '/simpleimage/2017/1113'))
 		self.assertTrue(os.path.isdir(self.OUTPUT_DIR + '/simplemask/2015/0208'))
